refactor: move schemas file #37
